pyproject.toml
[project]
name = "ro-agent"
version = "0.1.0"
description = "A read-only agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mysql-connector-python>=8.0",
    "openai>=2.14.0",
    "openpyxl>=3.1.0",
    "oracledb>=3.4.1",
    "prompt-toolkit>=3.0.52",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.21.0",
    "vertica-python>=1.4.0",
    "psycopg>=3.3.2",
    "cerebras-cloud-sdk>=1.64.1",
    "streamlit>=1.40",
    "pandas>=2.0",
]

[project.scripts]
ro-agent = "ro_agent.cli:app"
ro-eval = "ro_agent.eval.agentbench.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
# Harbor integration for TerminalBench evaluation
# Install harbor from: https://github.com/laude-institute/harbor
harbor = []  # Harbor is installed separately via pip install -e ~/proj/harbor

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
